+++ _event-store/src/adapters/amqp/emitter.rs
