+++ _event-store/src/adapters/amqp/mod.rs
