+++ _event-store/src/adapters/mod.rs
