+++ _event-store/src/adapters/stub/emitter.rs
