+++ _event-store/src/adapters/stub/mod.rs
