+++ _event-store/src/aggregator.rs
