+++ _event-store/src/store.rs
