+++ _event-store/src/store_query.rs
