+++ _event-store/src/utils.rs
