+++ _event-store/tests/aggregate.rs
