+++ event-store/Cargo.toml
@@ -3,31 +3,4 @@ name = "event-store"
 version = "0.1.0"
 authors = ["James Waples <jamwaffles@gmail.com>"]
 
-[badges]
-travis-ci = { repository = "repositive/event-store-rs", branch = "master" }
-
 [dependencies]
-serde = "1.0.80"
-serde_derive = "1.0.80"
-serde_json = "1.0.32"
-fallible-iterator = "0.1.5"
-sha2 = "0.8.0"
-chrono = { version = "0.4.6", features = ["serde"] }
-uuid = { version = "0.5.1", features = ["serde", "v4"] }
-event-store-derive = { path = "../event-store-derive" }
-event-store-derive-internals = { path = "../event-store-derive-internals" }
-log = "0.4.6"
-lapin-futures = "0.14.0"
-futures = "0.1.25"
-tokio = "0.1.11"
-r2d2 = "0.8.2"
-postgres = { version = "0.15.2", features = [ "with-serde_json", "with-uuid", "with-chrono" ] }
-r2d2_postgres = "0.14.0"
-
-[dev-dependencies]
-criterion = "0.2.5"
-env_logger = "0.5.13"
-
-[[bench]]
-name = "testhelpers"
-harness = false
